requirements.txt
black==22.10.0
Django==2.2.24
djangorestframework==3.9.0
drf-writable-nested==0.7.0
drf-yasg==1.2.1
gunicorn==20.1.0

